Move Table Toolbar to own component folder TableToolbar
