+++ src/TableToolbar/Toolbar.scss
@@ -1,4 +1,4 @@
-@import '../../common.scss';
+@import '../common.scss';
 
 .toolbar {
   display: flex;
