+++ stories/Table/TableExample.js
@@ -8,7 +8,7 @@ import {
   Title,
   SelectedCount,
   Divider
-} from 'wix-style-react/Table/Toolbar';
+} from 'wix-style-react/TableToolbar';
 
 import Card from 'wix-style-react/Card';
 import Search from 'wix-style-react/Search';
