+++ stories/TableToolbar/Toolbar.story.scss
@@ -1,4 +1,4 @@
-@import '../../../src/common.scss';
+@import '../../src/common.scss';
 
 .example {
   background-color: $D70;
